+++ extensions/msal-node-extensions/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @azure/msal-node-extensions
 
-<!-- This log was last generated on Wed, 05 Feb 2025 00:15:19 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 18 Feb 2025 18:51:41 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.5.5
+
+Tue, 18 Feb 2025 18:51:41 GMT
+
+### Patches
+
+- Bump @azure/msal-common to v15.2.0
+- Bump eslint-config-msal to v0.0.0
+- Bump rollup-msal to v0.0.0
 
 ## 1.5.4