+++ lib/msal-common/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @azure/msal-common
 
-<!-- This log was last generated on Wed, 05 Feb 2025 00:15:19 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 18 Feb 2025 18:51:40 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 15.2.0
+
+Tue, 18 Feb 2025 18:51:40 GMT
+
+### Minor changes
+
+- Pre-generate PKCE for async popup flow to blocked async popups in Safari #7568 (kshabelko@microsoft.com)
+- Bump eslint-config-msal to v0.0.0
+- Bump msal-test-utils to v0.0.1
+- Bump rollup-msal to v0.0.0
 
 ## 15.1.1