+++ lib/msal-react/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @azure/msal-react
 
-<!-- This log was last generated on Wed, 05 Feb 2025 00:15:19 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 18 Feb 2025 18:51:41 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 3.0.5
+
+Tue, 18 Feb 2025 18:51:41 GMT
+
+### Patches
+
+- Bump @azure/msal-browser to v4.3.0
+- Bump eslint-config-msal to v0.0.0
 
 ## 3.0.4